+++ php/tests/Application/Command/SemanticLintTest/GroupedUseStatements.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace B {
+    class Foo {
+
+    }
+
+    class Bar {
+
+    }
+}
+
+namespace A {
+    use \B\{
+        Foo,
+        Bar,
+        Missing
+    };
+}